requirements.txt
six==1.12.0
twisted==18.9.0
zope.interface==4.6.0
treq==18.6.0
feedparser==5.2.1
klein==17.10.0
lxml==4.3.3
attrs==19.1.0